requirements.txt
# Generated from pyproject.toml
pydantic>=2.0
pyyaml>=6.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.7.0
fastapi>=0.100.0
uvicorn>=0.20.0
sentence-transformers>=2.3.0
transformers>=4.35.0
torch>=2.0.0
faiss-cpu>=1.7.4
chromadb>=0.4.0
qdrant-client>=1.6.0
rank-bm25>=0.2.2
pyserini>=0.22.0
ragatouille>=0.0.7
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
dspy-ai>=2.0.0
llmlingua>=0.2.0
ragas>=0.1.0
trulens-eval>=0.25.0
datasets>=2.14.0
numpy>=1.24.0
tiktoken>=0.5.0
tqdm>=4.65.0
cleanlab>=2.5.0
